LayoutTests/imported/w3c/web-platform-tests/streams/readable-streams/async-iterator.any.sharedworker-expected.txt


FAIL Async iterator instances should have the correct list of properties assert_array_equals: should have all the correct methods lengths differ, expected array ["next", "return"] length 2, got ["next"] length 1
FAIL Async-iterating a push source assert_array_equals: lengths differ, expected array [1, 2, 3] length 3, got [] length 0
FAIL Async-iterating a pull source assert_array_equals: lengths differ, expected array [1, 2, 3] length 3, got [] length 0
FAIL Async-iterating a push source with undefined values assert_array_equals: lengths differ, expected array [undefined, undefined, undefined] length 3, got [] length 0
FAIL Async-iterating a pull source with undefined values assert_array_equals: lengths differ, expected array [undefined, undefined, undefined] length 3, got [] length 0
FAIL Async-iterating a pull source manually assert_equals: value expected (number) 1 but got (undefined) undefined
FAIL Async-iterating an errored stream throws assert_equals: expected (string) "e" but got (object) object "Error: assert_unreached: Reached unreachable code"
PASS Async-iterating a closed stream never executes the loop body, but works fine
FAIL Async-iterating an empty but not closed/errored stream never executes the loop body and stalls the async function assert_unreached: Reached unreachable code
FAIL Async-iterating a partially consumed stream assert_array_equals: lengths differ, expected array [2, 3] length 2, got [] length 0
FAIL Cancellation behavior when throwing inside loop body; preventCancel = false assert_array_equals: cancel() should be called lengths differ, expected array ["pull", "cancel", undefined] length 3, got [] length 0
FAIL Cancellation behavior when throwing inside loop body; preventCancel = true assert_array_equals: cancel() should not be called lengths differ, expected array ["pull"] length 1, got [] length 0
FAIL Cancellation behavior when breaking inside loop body; preventCancel = false assert_array_equals: cancel() should be called lengths differ, expected array ["pull", "cancel", undefined] length 3, got [] length 0
FAIL Cancellation behavior when breaking inside loop body; preventCancel = true assert_array_equals: cancel() should not be called lengths differ, expected array ["pull"] length 1, got [] length 0
FAIL Cancellation behavior when returning inside loop body; preventCancel = false assert_array_equals: cancel() should be called lengths differ, expected array ["pull", "cancel", undefined] length 3, got [] length 0
FAIL Cancellation behavior when returning inside loop body; preventCancel = true assert_array_equals: cancel() should not be called lengths differ, expected array ["pull"] length 1, got [] length 0
FAIL Cancellation behavior when manually calling return(); preventCancel = false promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return()', 'it.return' is undefined)"
FAIL Cancellation behavior when manually calling return(); preventCancel = true promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return()', 'it.return' is undefined)"
FAIL next() rejects if the stream errors assert_equals: 1st next() value expected (number) 0 but got (undefined) undefined
FAIL return() does not rejects if the stream has not errored yet promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value')', 'it.return' is undefined)"
FAIL return() rejects if the stream has errored promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value')', 'it.return' is undefined)"
FAIL next() that succeeds; next() that reports an error; next() assert_equals: 1st next() value expected (number) 0 but got (undefined) undefined
FAIL next() that succeeds; next() that reports an error(); next() [no awaiting] assert_equals: 1st next() value expected (number) 0 but got (undefined) undefined
FAIL next() that succeeds; next() that reports an error(); return() assert_equals: 1st next() value expected (number) 0 but got (undefined) undefined
FAIL next() that succeeds; next() that reports an error(); return() [no awaiting] promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value')', 'it.return' is undefined)"
FAIL next() that succeeds; return() assert_equals: next() value expected (number) 0 but got (undefined) undefined
FAIL next() that succeeds; return() [no awaiting] promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value')', 'it.return' is undefined)"
FAIL return(); next() promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value')', 'it.return' is undefined)"
FAIL return(); next() [no awaiting] promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value')', 'it.return' is undefined)"
FAIL return(); next() with delayed cancel() promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value')', 'it.return' is undefined)"
FAIL return(); next() with delayed cancel() [no awaiting] promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value')', 'it.return' is undefined)"
FAIL return(); return() promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value 1')', 'it.return' is undefined)"
FAIL return(); return() [no awaiting] promise_test: Unhandled rejection with value: object "TypeError: it.return is not a function. (In 'it.return('return value 1')', 'it.return' is undefined)"
FAIL values() throws if there's already a lock assert_throws_js: values() should throw function "() => s.values()" did not throw
FAIL Acquiring a reader after exhaustively async-iterating a stream assert_array_equals: lengths differ, expected array [1, 2, 3] length 3, got [] length 0
FAIL Acquiring a reader after return()ing from a stream that errors assert_equals: 1st next() value expected (number) 0 but got (undefined) undefined
FAIL Acquiring a reader after partially async-iterating a stream assert_array_equals: lengths differ, expected array [1, 2] length 2, got [] length 0
FAIL Acquiring a reader and reading the remaining chunks after partially async-iterating a stream with preventCancel = true assert_array_equals: lengths differ, expected array [1, 2] length 2, got [] length 0
FAIL return() should unlock the stream synchronously when preventCancel = false rs.values({ preventCancel }).return is not a function. (In 'rs.values({ preventCancel }).return()', 'rs.values({ preventCancel }).return' is undefined)
FAIL return() should unlock the stream synchronously when preventCancel = true rs.values({ preventCancel }).return is not a function. (In 'rs.values({ preventCancel }).return()', 'rs.values({ preventCancel }).return' is undefined)
FAIL close() while next() is pending assert_array_equals: lengths differ, expected array ["a", "b", "c"] length 3, got [] length 0
